streamlit-app/requirements.txt
google-auth-oauthlib==1.0.0
google-api-python-client==2.88.0
google-search-results==2.4.2
langchain==0.0.196
numpy==1.24.3
openai==0.27.8
redis==4.5.5
streamlit==1.23.1
streamlit-chat==0.0.2.2
streamlit-extras==0.2.7
torch==2.0.1
transformers==4.30.1
watchdog==3.0.0